Fixed asprintf usage

+++ src/xine-utils/xmlparser.c
@@ -637,7 +637,7 @@ static void xml_parser_dump_node (const xml_node_t *node, int indent) {
     free (value);
     p = p->next;
     if (p) {
-      printf ("\n%*s", indent+2+l, "");
+      printf ("\n%*s", (int)(indent+2+l), "");
     }
   }
   printf (">\n");
